+++ src/glamps/views.py
@@ -1,3 +1,5 @@
+from django.db import transaction
+from django.shortcuts import get_object_or_404
 from drf_spectacular.utils import extend_schema
 from rest_framework import status
 from rest_framework.response import Response
@@ -200,16 +202,19 @@ class GlampModelViewSet(ModelViewSet, EventLogMixin):
         serializer = self.get_serializer(queryset, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
+    @transaction.atomic
     def perform_create(self, serializer):
         glamp_instance = serializer.save()
         self.log_event(self.request, glamp_instance)
         return glamp_instance
 
+    @transaction.atomic
     def perform_update(self, serializer):
         glamp_instance = serializer.save()
         self.log_event(self.request, glamp_instance)
         return glamp_instance
 
+    @transaction.atomic
     def perform_destroy(self, glamp_instance):
         self.log_event(self.request, glamp_instance)
         return super().perform_destroy(glamp_instance)
@@ -224,6 +229,7 @@ class GlampByCategoryViewSet(ModelViewSet, EventLogMixin):
 
     def get_queryset(self):
         category_id = self.kwargs.get("category_id")
+        get_object_or_404(Category, id=category_id)
         return Glamp.objects.filter(category_id=category_id)
 
     def get_permissions(self):
@@ -247,18 +253,22 @@ class GlampByCategoryViewSet(ModelViewSet, EventLogMixin):
 
         return [permission() for permission in permission_classes]
 
+    @transaction.atomic
     def perform_create(self, serializer):
         category_id = self.kwargs.get("category_id")
-        category = Category.objects.get(id=category_id)
+        category = get_object_or_404(Category, id=category_id)
         glamp_instance = serializer.save(category=category)
         self.log_event(self.request, glamp_instance)
         return glamp_instance
 
+    @transaction.atomic
     def perform_update(self, serializer):
         glamp_instance = serializer.save()
         self.log_event(self.request, glamp_instance)
         return glamp_instance
 
+    @transaction.atomic
     def perform_destroy(self, glamp_instance):
         self.log_event(self.request, glamp_instance)
         return super().perform_destroy(glamp_instance)
+